RST and ProgMem initialization
The previous instruction memory did not have a rst execution, so it has been added. Also, now the testbench should have always a rst cycle to initialize the ProgMem in order to avoid glitches like not executing the first test o executing always the first instruction from the previous test.