Core Instruction Expansion set M last update
The division module has been replaced by the "double skip execution" (dseDIVrest32) division design. This is high on resource consumption (check DIVrest folder for exact resource usage in the Cyclone III D0 FPGA platform synthesis) but is the best performing with a maximum clock latency of 18 clock cycles per division operation.

The multiplication module has been replaced by a proper IP, the LPM_MULT IP being specific. Check .comp_sv file to learn new details to take into account when performing simulations. Also, now is required to have a Quartus or at least the LPM libraries to be able to compile the project. In any case, the MULgold module has been left in the files, just in case this libraries are exempt, so in order to compile the project without the LPM libraries, the MUL module has to be changed in the core_execution_unit_alu_m.v file to the MULgold and comment the LPM_MULT32 that is being used by default.