Core project with memories
Given that the memories have been modified to send a NOP instruction when there's a reset, it has been opted to pack the core, data and instruction memories within the same package.

This means that now the renamed "top_CoreMem.v" is a a standalone module capable of executing and programing through the named "AXI" bus. There're some input and wire un/commented because without them, it's impossible to simulate properly, so make sure to un/comment to be able program the instruction memory from outside of the module.