Resource usage estimation and MULDIV2
It has been updated the resource usage estimation with the information of all DIV, MUL and MULDIV modules, with the addition of the maximum frequency of operation specified at the "MULDIV performance.txt" file.

A MULDIV2 design has been uploaded that differs with the MULDIV in that uses the ALTMULT_ADD multiplication module IP, which has different frequency performance.

Some RTL mapping are still missing.