RV32IMF tested
Many errors have been corrected at the FPU and its modules. The FPU is also added to the RV32IM core with some modifications to make it suitable for fused operations (that are not implemented yet, but prepared).

The only thing structural that is lacking is the FCSR, that I'm unable to understand properly about how the Ziscr instruction set works.