C2RISC-V guide and debug performed
A C program to RISC-V machine code tutorial has been written. It's nothing advanced, but it saves hours researching to start on programing software.

This has been used to code a little C program with adds and a for. After putting it's RISC-V conversion code, the cores just didn't work because of a number of bugs. These have been addressed and solved.

The RV32IMF is the only core tested with this, and it passes all the previous tests. The changes have been ported by hand to the other RV cores (RV32IM and RV32I. the non-pipeline version has been discontinuated).