IP Divider comparison
Well, due to some questions during the thesis presentation (9.6 btw), I felt as a requirement to generate LPM_DIVIDE IP modules from the IP library to compare with the dividers we have developed.

This update justs adds these IP modules and the comparison.

In summary, the IP library generates pipeline dividers that require many resources for a characteristic that is not taken advantatge from at our core, making a waste in space.
So in general, any of our dividers are better than the IPs. However, the realization has been made about designing an Quad and Octo divider using Carry-LookAhead adders, but that will be left undone most probably.